requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via pgqueuer
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via
    #   pgcachewatch
    #   pgqueuer
idna==3.7
    # via anyio
pgcachewatch==0.4.2
    # via pgqueuer
pgqueuer==0.3.4
    # via -r requirements.in
pydantic==2.7.1
    # via
    #   pgcachewatch
    #   pgqueuer
pydantic-core==2.18.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
tabulate==0.9.0
    # via pgqueuer
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
websockets==12.0
    # via pgcachewatch
